+++ core/executor/src/debugger/create2.rs
@@ -26,7 +26,7 @@ async fn test_create2_gas() {
     // );
     // println!("{:?}", resp);
 
-    let gas_used = resp.gas_used;
+    let gas_used = resp.tx_resp[0].gas_used;
     let after_balance = debugger.backend(1).basic(sender).balance;
 
     assert_eq!(gas_used * 8, (init_balance - after_balance).as_u64());
