+++ core/executor/src/lib.rs
@@ -13,9 +13,9 @@ mod vm;
 
 pub use crate::adapter::{AxonExecutorAdapter, MPTTrie, RocksTrieDB};
 pub use crate::system::NATIVE_TOKEN_ISSUE_ADDRESS;
-pub use crate::utils::{decode_revert_msg, logs_bloom};
+pub use crate::utils::{code_address, decode_revert_msg, logs_bloom};
 pub use crate::vm::{
-    code_address, CROSSCHAIN_CONTRACT_ADDRESS, METADATA_CONTRACT_ADDRESS, WCKB_CONTRACT_ADDRESS,
+    CROSSCHAIN_CONTRACT_ADDRESS, METADATA_CONTRACT_ADDRESS, WCKB_CONTRACT_ADDRESS,
 };
 
 use std::collections::BTreeMap;
