+++ core/executor/src/vm.rs
@@ -3,9 +3,7 @@ use std::collections::BTreeMap;
 use evm::executor::stack::{MemoryStackState, PrecompileFn, StackExecutor, StackSubstateMetadata};
 
 use protocol::traits::{ApplyBackend, Backend};
-use protocol::types::{
-    Config, Hasher, SignedTransaction, TransactionAction, TxResp, H160, H256, U256,
-};
+use protocol::types::{Config, SignedTransaction, TransactionAction, TxResp, H160};
 
 pub const METADATA_CONTRACT_ADDRESS: H160 = H160([
     161, 55, 99, 105, 25, 112, 217, 55, 61, 79, 171, 124, 195, 35, 215, 186, 6, 250, 153, 134,
@@ -16,7 +14,6 @@ pub const WCKB_CONTRACT_ADDRESS: H160 = H160([
 pub const CROSSCHAIN_CONTRACT_ADDRESS: H160 = H160([
     246, 123, 196, 229, 13, 29, 249, 43, 14, 76, 97, 121, 74, 69, 23, 175, 106, 153, 92, 178,
 ]);
-const RETURN_MESSAGE_PRUNE_PREFIX: usize = 4;
 
 // deprecated
 #[allow(dead_code)]
@@ -73,7 +70,7 @@ impl EvmExecutor {
         let code_address = if tx.transaction.unsigned.action() == &TransactionAction::Create
             && exit_reason.is_succeed()
         {
-            Some(code_address(&tx.sender, &old_nonce))
+            Some(crate::code_address(&tx.sender, &old_nonce))
         } else {
             None
         };
@@ -94,61 +91,5 @@ impl EvmExecutor {
         }
 
         resp
-    }
-}
-
-pub fn code_address(sender: &H160, nonce: &U256) -> H256 {
-    let mut stream = rlp::RlpStream::new_list(2);
-    stream.append(sender);
-    stream.append(nonce);
-    Hasher::digest(&stream.out())
-}
-
-pub fn decode_revert_msg(input: &[u8]) -> String {
-    String::from_iter(
-        input
-            .iter()
-            .filter_map(|i| {
-                let c = *i as char;
-                if c.is_control() {
-                    None
-                } else {
-                    Some(c)
-                }
-            })
-            .skip(RETURN_MESSAGE_PRUNE_PREFIX),
-    )
-}
-
-#[cfg(test)]
-mod tests {
-    use protocol::codec::{hex_decode, hex_encode};
-
-    use super::*;
-
-    #[test]
-    fn test_code_address() {
-        let sender = H160::from_slice(
-            hex_decode("8ab0cf264df99d83525e9e11c7e4db01558ae1b1")
-                .unwrap()
-                .as_ref(),
-        );
-        let nonce: U256 = 0u64.into();
-        let addr: H160 = code_address(&sender, &nonce).into();
-        assert_eq!(
-            hex_encode(addr.0).as_str(),
-            "a13763691970d9373d4fab7cc323d7ba06fa9986"
-        );
-
-        let sender = H160::from_slice(
-            hex_decode("6ac7ea33f8831ea9dcc53393aaa88b25a785dbf0")
-                .unwrap()
-                .as_ref(),
-        );
-        let addr: H160 = code_address(&sender, &nonce).into();
-        assert_eq!(
-            hex_encode(addr.0).as_str(),
-            "cd234a471b72ba2f1ccf0a70fcaba648a5eecd8d"
-        )
     }
 }
